test: end to end test fashion single file per swap type
