+++ .github/workflows/CI.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Start Bitcoin Core container
+      - name: Start Elements Core container
         run: npm run docker:start
 
       - name: Compile
