+++ test-lib/unit/swap/Utils.ts
@@ -3,6 +3,9 @@ import { OutputType } from '../../../lib/consts/Enums';
 import { p2wshOutput, p2shOutput, p2shP2wshOutput } from '../../../lib/swap/Scripts';
 import { PrefixUnconfidential } from '../../../lib/consts/Buffer';
 import { networks } from 'liquidjs-lib';
+import ECPairFactory, { ECPairAPI, TinySecp256k1Interface } from 'ecpair';
+import * as ecc from 'tiny-secp256k1';
+
 
 export const getScriptHashFunction = (type: OutputType): (scriptHex: Buffer) => Buffer => {
   switch (type) {
@@ -17,3 +20,6 @@ export const LBTC_REGTEST = Buffer.concat([
   PrefixUnconfidential,
   Buffer.from(networks.regtest.assetHash, 'hex').reverse(),
 ]);
+
+const tinysecp: TinySecp256k1Interface = ecc;
+export const ECPair: ECPairAPI = ECPairFactory(tinysecp);
